requirements.txt
# Requirements for captcha-bot
# Add your dependencies here
# Example:
discord.py>=2.0.0
Pillow
captcha